move exceptions around
