+++ wandb/apis/internal_runqueue.py
@@ -16,7 +16,7 @@ from six import BytesIO
 import wandb
 from wandb import __version__, env, util
 from wandb.apis.normalize import normalize_exceptions
-from wandb.errors.error import CommError, UsageError
+from wandb.errors import CommError, UsageError
 from wandb.old import retry
 from wandb.old.settings import Settings
 import yaml
