+++ wandb/apis/normalize.py
@@ -11,7 +11,7 @@ from gql.client import RetryError
 import requests
 import six
 from wandb import env
-from wandb.errors.error import CommError
+from wandb.errors import CommError
 
 
 def normalize_exceptions(func):
