+++ wandb/sdk_py27/internal/internal_api.py
@@ -27,7 +27,7 @@ from wandb.old.settings import Settings
 from wandb.old import retry
 from wandb import util
 from wandb.apis.normalize import normalize_exceptions
-from wandb.errors.error import CommError, UsageError
+from wandb.errors import CommError, UsageError
 from ..lib.filenames import DIFF_FNAME
 from ..lib.git import GitRepo
