+++ wandb/sdk_py27/wandb_init.py
@@ -20,7 +20,7 @@ import shortuuid  # type: ignore
 import six
 import wandb
 from wandb import trigger
-from wandb.errors.error import UsageError
+from wandb.errors import UsageError
 from wandb.integration import sagemaker
 from wandb.integration.magic import magic_install
 from wandb.util import sentry_exc
